hector_quadrotor_model: make models more robust against irregular input values
Especially on collisions with walls and obstacles Gazebo can output very high
velocity values for some reason, which caused the propulsion and aerodynamic
models to output infinite values or NaN as resulting forces and torques, with
the consequence that Gazebo effectively stopped simulation and showed the quadrotor
at the origin of the world frame.

With this patch the simulation is much more stable in case of collisions or
hard landings.